parser/pyproject.toml
[project]
name = "read-ais"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
    "geopandas>=1.1.2",
]

[project.scripts]
read-ais = "read_ais:main"

[build-system]
requires = ["uv_build>=0.9.10,<0.10.0"]
build-backend = "uv_build"
